From 5d32ec3e897220765cff85387708632de28a14ff Mon Sep 17 00:00:00 2001 From: Perry Fraser <perry@frasers.org> Date: Mon, 20 Jun 2022 17:39:09 -0400 Subject: [PATCH] Add style exception for OOUI buttons' margin
On OOUI forms the text for the checkboxes tended to be rather
misaligned, with the text usually ending up getting shunted off to the
bottom-right corner. It seems the margin being added by the special-case
in this rule was responsible, so I added an exception to it. In testing
things seem to look okay.

Change-Id: Id3ca935e5c83d7c96edcb02d1c6fd72af8e81e58